fix(script): fixed all remaining bugs
Fixed the percentage calculation to the result of the previous operation.
Set 0 as the first number by default, so as not to create errors if a sign is pressed first rather than a number.
Fixed the ability to change sign before selecting a second number.
Added rounding to the solution, including the first two decimals.
Adjusted negation to the result of the previous operation instead of the second number selected.