fix(script): fixed small bugs
Fixed the equal key in case it should be pressed before the mark. Now if you press it as a result the first number that was selected will come up.
Fixed the 0 key, if it should be pressed when there is only a 0, no more zeros will appear.
Fixed the selection of a number after doing a calculation. Now if the previous calculation is pressed, it is deleted, and only the selected number appears.